+++ fabrictestbed_extensions/ui/artifact_manager_ui.py
@@ -23,11 +23,14 @@
 #
 # Author: Komal Thareja(kthare10@renci.org)
 
+import os
 import tarfile
+
 import ipywidgets as widgets
 from IPython.display import display
+
+from fabrictestbed_extensions.fablib.constants import Constants
 from fabrictestbed_extensions.fablib.fablib import FablibManager
-import os
 
 
 class ArtifactManagerUI:
@@ -46,11 +49,15 @@ class ArtifactManagerUI:
         Widget to display error messages.
     title_filter : widgets.Text
         Text input widget for filtering artifacts by title.
+    tag_filter : widgets.Text
+        Text input widget for filtering artifacts by tags.
+    project_filter : widgets.Text
+        Text input widget for filtering artifacts by project.
     grid : widgets.GridBox
         GridBox widget displaying the artifact information.
     """
 
-    def __init__(self, fablib: FablibManager, download_dir: str="/home/fabric/work"):
+    def __init__(self, fablib: FablibManager, download_dir: str = "/home/fabric/work"):
         """
         Initializes the ArtifactManagerUI with a specified download directory.
 
@@ -63,17 +70,28 @@ class ArtifactManagerUI:
         self.fablib = fablib
         self.artifacts = self.fablib.get_manager().list_artifacts()
         self.error_output = widgets.Output()
+
+        # Initialize title and tag filters
         self.title_filter = widgets.Text(
-            description='Filter by Title:',
-            placeholder='Enter title keyword',
-            layout=widgets.Layout(width='100%')
+            description="Title:",
+            placeholder="Enter title keyword",
+            layout=widgets.Layout(width="50%"),
         )
-        self.title_filter.observe(self.filter_artifacts, names='value')
+        self.tag_filter = widgets.Text(
+            description="Tag:",
+            placeholder="Enter tag keyword",
+            layout=widgets.Layout(width="50%"),
+        )
+
+        self.title_filter.observe(self.filter_artifacts, names="value")
+        self.tag_filter.observe(self.filter_artifacts, names="value")
 
         # Initialize grid as None
         self.grid = None
 
-    def download_artifact_handler(self, artifact_title: str, version_urn: str, version: str):
+    def download_artifact_handler(
+        self, artifact_title: str, version_urn: str, version: str
+    ):
         """
         Handler for downloading and extracting artifacts.
 
@@ -94,9 +112,14 @@ class ArtifactManagerUI:
         with self.error_output:
             try:
                 self.error_output.clear_output()  # Clear any previous error messages
-                location = self.fablib.download_artifact(download_dir=self.download_dir,
-                                                         version_urn=version_urn, version=version)
-                print(f"Artifact '{artifact_title}' version: '{version}' downloaded to {location} successfully!")
+                location = self.fablib.download_artifact(
+                    download_dir=self.download_dir,
+                    version_urn=version_urn,
+                    version=version,
+                )
+                print(
+                    f"Artifact '{artifact_title}' version: '{version}' downloaded to {location} successfully!"
+                )
 
                 # Extract the tar file at the same location
                 if tarfile.is_tarfile(location):
@@ -104,7 +127,8 @@ class ArtifactManagerUI:
                         extract_path = os.path.dirname(location)
                         tar.extractall(path=extract_path)
                         print(
-                            f"Artifact '{artifact_title}' version: '{version}' extracted to {extract_path} successfully!")
+                            f"Artifact '{artifact_title}' version: '{version}' extracted to {extract_path} successfully!"
+                        )
                 else:
                     print(f"Downloaded file '{location}' is not a valid tar file.")
 
@@ -113,7 +137,7 @@ class ArtifactManagerUI:
 
     def filter_artifacts(self, change):
         """
-        Filters artifacts based on the title filter text input.
+        Filters artifacts based on the title, tag or project filter text input.
 
         Parameters
         ----------
@@ -121,8 +145,18 @@ class ArtifactManagerUI:
             The change dictionary containing the new value of the text input.
         """
         filter_text = self.title_filter.value.lower()
-        filtered_artifacts = [artifact for artifact in self.artifacts
-                              if filter_text in artifact['title'].lower()]
+        filtered_artifacts = [
+            artifact
+            for artifact in self.artifacts
+            if filter_text in artifact["title"].lower()
+        ]
+        tag_text = self.tag_filter.value.lower()
+        filtered_artifacts = [
+            artifact
+            for artifact in filtered_artifacts
+            if tag_text in ", ".join(artifact["tags"])
+        ]
+
         self.update_ui(filtered_artifacts)
 
     def update_ui(self, artifacts):
@@ -136,62 +170,95 @@ class ArtifactManagerUI:
         """
         # Clear existing grid content, if any
         if self.grid is not None:
-            self.grid.children = []  # Clear the grid's content without removing the widget itself
+            self.grid.children = (
+                []
+            )  # Clear the grid's content without removing the widget itself
 
         # Define table headers
         headers = [
             widgets.HTML(value="<b>Title</b>"),
-            widgets.HTML(value="<b>Authors</b>"),
             widgets.HTML(value="<b>Description</b>"),
             widgets.HTML(value="<b>Tags</b>"),
             widgets.HTML(value="<b>Versions</b>"),
+            widgets.HTML(value="<b>Download</b>"),
+            widgets.HTML(value="<b>Project</b>"),
+            widgets.HTML(value="<b>Authors</b>"),
         ]
+
+        for h in headers:
+            h.style.text_align = "center"
+            h.style.text_color = Constants.FABRIC_WHITE
+            h.style.background = Constants.FABRIC_PRIMARY
 
         # Prepare grid items
         grid_items = headers.copy()
 
         # Build interactive UI for each artifact
         for artifact in artifacts:
-            title_label = widgets.HTML(value=artifact['title'])
-            authors_label = widgets.HTML(value=', '.join([author['name'] for author in artifact['authors']]))
-            description_label = widgets.HTML(value=artifact['description_short'])
-            tags_label = widgets.HTML(value=', '.join(artifact['tags']))
+            title_label = widgets.HTML(value=artifact["title"])
+            authors_label = widgets.HTML(
+                value=", ".join([author["name"] for author in artifact["authors"]])
+            )
+            description_label = widgets.HTML(value=artifact["description_short"])
+            tags_label = widgets.HTML(value=", ".join(artifact["tags"]))
+            project_label = widgets.HTML(value=artifact.get("project_name", "N/A"))
 
-            # Version buttons in a horizontal box
-            version_buttons = []
-            for version in artifact["versions"]:
-                button = widgets.Button(
-                    description=f"Download v{version['version']}",
-                    tooltip=f"Download version {version['version']}",
-                    button_style='success',
-                    layout=widgets.Layout(width='auto')
+            # Version dropdown list
+            version_options = [
+                (f"v{version['version']}", version["urn"])
+                for version in artifact["versions"]
+            ]
+            version_dropdown = widgets.Dropdown(
+                options=version_options,
+                description="",
+                layout=widgets.Layout(width="auto"),
+            )
+
+            # Download button
+            download_button = widgets.Button(
+                description="Download",
+                tooltip="Download selected version",
+                layout=widgets.Layout(width="auto"),
+            )
+            download_button.style.button_color = Constants.FABRIC_PRIMARY_LIGHT
+
+            # Bind the download handler
+            download_button.on_click(
+                lambda b, artifact_title=artifact[
+                    "title"
+                ], version_dropdown=version_dropdown: self.download_artifact_handler(
+                    artifact_title, version_dropdown.value, version_dropdown.label
                 )
-                # Ensure unique binding for each button
-                button.on_click(lambda b, artifact_title=artifact['title'], version_urn=version["urn"],
-                                       version=version['version']: self.download_artifact_handler(artifact_title,
-                                                                                                  version_urn, version))
-                version_buttons.append(button)
+            )
 
-            versions_box = widgets.HBox(version_buttons)
-
-            # Append artifact details and buttons to grid items
-            grid_items.extend([title_label, authors_label, description_label, tags_label, versions_box])
+            # Append artifact details and controls to grid items
+            grid_items.extend(
+                [
+                    title_label,
+                    description_label,
+                    tags_label,
+                    version_dropdown,
+                    download_button,
+                    project_label,
+                    authors_label,
+                ]
+            )
 
         # Create and display the grid layout
         if self.grid is None:
             self.grid = widgets.GridBox(
                 grid_items,
                 layout=widgets.Layout(
-                    grid_template_columns="repeat(5, 25%)",
+                    grid_template_columns="20% 30% 10% 10% 10% 10% 10%",
                     grid_gap="10px",
-                    align_items="center"
-                )
+                    align_items="center",
+                ),
             )
         else:
             self.grid.children = grid_items
 
         # Display the filter and the grid
-        display(self.title_filter)
+        display(widgets.HBox([self.title_filter, self.tag_filter]))
         display(self.grid)
         display(self.error_output)
 
@@ -202,7 +269,7 @@ class ArtifactManagerUI:
         self.update_ui(self.artifacts)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     # Create an instance of the class and render the UI
     artifact_manager_ui = ArtifactManagerUI(fablib=FablibManager())
     artifact_manager_ui.create_ui()